requirements.txt
chainlit>=1.0.0
azure-ai-projects
azure-identity
